requirements.txt
fastapi[all] 
uvicorn
pydantic
sqlalchemy
psycopg2-binary
watchfiles
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
orjson
httpx
pgvector
apscheduler


# ✅ deepresearch 관련
firecrawl
annotated-types
anyio
certifi==2025.1.31
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
idna==3.10
jiter==0.8.2
openai==1.64.0
pydantic_core==2.27.2
sniffio==1.3.1
typing_extensions==4.12.2


# ✅ LangChain 및 벡터 DB 관련
langchain-core
langchain
langchain-community
langchain-openai
tavily-python
elasticsearch[async]
kiwipiepy

# ✅ NLP 및 모델 관련
transformers
pandas

# ✅ 벡터 DB 관련
faiss-cpu
